refactor: infrastructure 패키지 삭제
