+++ src/main/java/kr/co/fastcampus/travel/service/ItineraryService.java
@@ -7,7 +7,7 @@ import kr.co.fastcampus.travel.common.exception.TravelDoesNotExistException;
 import kr.co.fastcampus.travel.controller.dto.ItinerarySaveRequest;
 import kr.co.fastcampus.travel.domain.Itinerary;
 import kr.co.fastcampus.travel.domain.Trip;
-import kr.co.fastcampus.travel.infrastructure.repository.ItineraryRepository;
+import kr.co.fastcampus.travel.repository.ItineraryRepository;
 import kr.co.fastcampus.travel.view.enums.FileType;
 import lombok.RequiredArgsConstructor;
